extract queries to another distinct folder
